requirements.txt
slackclient==2.9.3
selenium==3.141.0
loguru==0.5.3
prettytable==2.1.0